refactor: Extract shared recovery types and split restore paths
- Move RecoveredState into recovery/types.rs for shared use
- Extract full_journal_replay into recovery/journal_restore.rs
- Split ExecutionConfig into its own execution_config.rs file
