+++ stepflow-rs/crates/stepflow-plugin/src/execution_config.rs
@@ -1,0 +1,25 @@
+// Copyright 2025 DataStax Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
+// in compliance with the License. You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software distributed under the License
+// is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+// or implied. See the License for the specific language governing permissions and limitations under
+// the License.
+
+//! Execution configuration stored in the environment.
+
+/// Execution configuration stored in the environment.
+///
+/// Consolidates runtime execution options that affect the executor.
+#[derive(Debug, Clone)]
+pub struct ExecutionConfig {
+    /// Number of journal entries between checkpoints.
+    ///
+    /// When non-zero, the executor creates a checkpoint every N journal entries.
+    /// Set to 0 to disable. Default: 0 (builder), 1000 (config file).
+    pub checkpoint_interval: usize,
+}
