+++ stepflow-rs/crates/stepflow-plugin/src/lib.rs
@@ -14,14 +14,16 @@ mod context;
 mod environment_builder;
 mod environment_ext;
 mod error;
+mod execution_config;
 mod plugin;
 pub mod routing;
 mod subflow;
 
 pub use context::RunContext;
-pub use environment_builder::{BlobApiUrl, ExecutionConfig, StepflowEnvironmentBuilder};
+pub use environment_builder::{BlobApiUrl, StepflowEnvironmentBuilder};
 pub use environment_ext::PluginRouterExt;
 pub use error::{PluginError, Result};
+pub use execution_config::ExecutionConfig;
 pub use plugin::{DynPlugin, Plugin, PluginConfig};
 pub use subflow::{
     SubflowReceiver, SubflowRequest, SubflowSubmitError, SubflowSubmitter, subflow_channel,
